libs/infinity_emb/pyproject.toml
[tool.poetry]
name = "infinity_emb"
version = "0.0.51"
description = "Infinity is a high-throughput, low-latency REST API for serving vector embeddings, supporting a wide range of sentence-transformer models and frameworks."
authors = ["michaelfeil <me@michaelfeil.eu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "infinity_emb"}]
homepage = "https://github.com/michaelfeil/infinity"
repository = "https://github.com/michaelfeil/infinity"
keywords = ["vector", "embedding", "neural", "search", "sentence-transformers"]

[tool.poetry.dependencies]
python = ">=3.9,<4"
# basics, pin numpy <2 for onnx
numpy = ">=1.20.0,<2"
huggingface_hub="*"
# logging
rich = {version = "^13", optional=true}
# webserver-only
fastapi = {version = ">=0.103.2", optional=true}
orjson = {version = ">=3.9.8,!=3.10.0", optional=true} 
prometheus-fastapi-instrumentator = {version = ">=6.1.0", optional=true}
uvicorn = {version = "^0.23.2", optional=true, extras = ["standard"]}
typer = {version = "^0.9.0", optional=true, extras = ["all"]}
pydantic = {version = ">=2.4.0,<3", optional=true}  
# backend
torch = {version = ">=2.2.1", source = "pypi", optional=true}
sentence-transformers = {version = "^3.0.1", optional=true} 
transformers = {version = ">4.34.0,<=5.0", optional=true} 
ctranslate2 = {version = "^4.0.0", optional=true}
optimum = {version = ">=1.16.2", optional=true, extras=["onnxruntime"]}
hf_transfer = {version=">=0.1.5"}
einops = {version = "*", optional=true}
# vision
pillow = {version = "*", optional=true}
timm = {version = "*", optional=true}
# cache
diskcache = {version = "*", optional=true}
# gpu
onnxruntime-gpu = {version = "*", optional=true} 
tensorrt = {version = "^8.6.1", optional=true}
opentelemetry-instrumentation-logging = "^0.46b0"
starlette = "^0.37.2"
opentelemetry-instrumentation = "^0.46b0"
opentelemetry-instrumentation-asgi = "^0.46b0"

[tool.poetry.scripts]
infinity_emb = "infinity_emb.infinity_server:cli"

[tool.poetry.group.test.dependencies]
pytest = "^7.0.0"
pytest-mock = "*"
httpx  = "*"
asgi_lifespan = "*"
anyio = "*"
trio = "*"
coverage = {extras = ["toml"], version = "^7.3.2"}
mypy = "^1.5.1"

# preferred dev dependencies
torch = "2.3.1"
prometheus-fastapi-instrumentator = "7.0.0"
fastapi = "0.110.2"

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.0"

[tool.poetry.group.mkdocs.dependencies]
mkdocs-material = "*"
mike = "*"
mkdocs-swagger-ui-tag = "*"

[tool.poetry.group.lint.dependencies]
ruff = "^0.2.0"
types-toml = "^0.10.8.1"
types-redis = "^4.3.21.6"
types-pytz = "^2023.3.0.0"
black = "^23.1.0"
types-chardet = "^5.0.4.6"
mypy-protobuf = "^3.0.0"

[tool.poetry.extras]
ct2=["ctranslate2","sentence-transformers","torch","transformers"]
optimum=["optimum"]
torch=["sentence-transformers","torch"]
einops=["einops"]
logging=["rich"]
cache=["diskcache"]
vision=["pillow","timm"]
server=[
    "fastapi",
    "orjson",
    "prometheus-fastapi-instrumentator",
    "pydantic",
    "rich",
    "typer",
    "uvicorn",
]
all=[
    "ctranslate2",
    "diskcache",
    "einops",
    "fastapi", 
    "optimum",
    "orjson", 
    "pillow",
    "prometheus-fastapi-instrumentator", 
    "pydantic", 
    "rich", 
    "sentence-transformers",
    "timm",
    "torch", 
    "typer", 
    "uvicorn",
]
# non-default gpu
tensorrt=["tensorrt"]
onnxruntime-gpu=["onnxruntime-gpu"]

[tool.pytest.ini_options]
markers = [
    "performance: tests that measure performance (deselect with '-m \"not performance\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
